Noveno Commit - Jorge

+++ src/java/Dao/PersonaDao.java
@@ -8,6 +8,6 @@ public interface PersonaDao {
     public boolean agregarPersona(Persona persona);
     public List<Persona> listarPersona();
     public boolean actualizarPersona(Persona persona);
-    public Persona buscarPersonaId(long id_persona);
-    public boolean eliminarPersona(long id_persona);
+    public Persona buscarPersonaId(int id_persona);
+    public boolean eliminarPersona(int id_persona);
 }
